SGA-90 Fix ConcurrentModificationException
The spurious ConcurrentModificationExceptions stem from not locking
the synchronized objects when iterating over them.
Javadoc makes it quite clear:

"It is imperative that the user manually synchronize on the returned
list when iterating over it"
https://docs.oracle.com/javase/7/docs/api/java/util/Collections.html
 #synchronizedList(java.util.List)
 #synchronizedMap(java.util.Map)
 #synchronizedSet(java.util.Set)

Thus an individual modification (add/remove/..) is synchronized but
the iteration is not! When a modification happens while iterating
an exception will be thrown instead of going into undefined behavior
as is the case without synchronizedLists/Maps/Sets

This commit fixes all such instances in the app

Change-Id: I638ffb02332a833d435d55905233282fd216f3a3