SGA-92 Replace libble 1.x with new libSmartGadget and libble 2.0
ScanDeviceFragment and ManageDeviceFragment are rewritten.
Thus, the Manage Device Screen, which is shown when clicking on a
connected gadget in the ScanDeviceFragment is also redesigned.

The old libble 1.x is entirely removed and replaced with
the new libsmartgadget and the new libble 2.

The back button bug SGA-77 in Manage Device Fragment is also fixed

Change-Id: Ib082410afd3eaaefaffc2260523de2be66952f74