Fix aggregation of RH and T values for download
When downloading logged RH/T values, these do not arrive in pairs,
but independently. Also, they arrive in packets, so several RH values
might arrive before the corresponding T value does. The code to
aggregate these values in pairs did not account for this, losing
three out of four values by overwriting them in the aggregator.

There are now two different aggregation strategies for live and
logged data: for live values, the timestamp is ignored, for historic
values it is not.